requirements.txt
Django==3.0.6
django-crispy-forms==1.9.0
django-taggit==1.3.0
Pillow==7.1.1
pytz==2020.1
